+++ BioViewer/Metal/Meshes/CreateImpostorSpheres.metal
@@ -15,7 +15,7 @@ using namespace metal;
 // MARK: - Kernels
 
 kernel void createImpostorSpheres(const device simd_float3 *atomPoints [[ buffer(0) ]],
-                                  const device uint8_t *atomType [[ buffer(1) ]],
+                                  const device uint16_t *atomType [[ buffer(1) ]],
                                   device BillboardVertex *generatedVertices [[ buffer(2) ]],
                                   device uint32_t *generatedIndices [[ buffer(3) ]],
                                   constant uint32_t & totalAtomCount [[ buffer(4) ]],
